+++ javadoc/index-files/index-16.html
@@ -171,7 +171,7 @@ loadScripts(document, 'script');</script>
 <div class="block">Game Logic to calculate reinforcement armies for each player based on
  number of countries captured.</div>
 </dd>
-<dt><span class="member-name-link"><a href="../model/GameSettings.html#setStrategy(model.strategy.GameStrategy)">setStrategy(GameStrategy)</a></span> - Method in class model.<a href="../model/GameSettings.html" title="class in model">GameSettings</a></dt>
+<dt><span class="member-name-link"><a href="../model/GameSettings.html#setStrategy(model.strategy.game.GameStrategy)">setStrategy(GameStrategy)</a></span> - Method in class model.<a href="../model/GameSettings.html" title="class in model">GameSettings</a></dt>
 <dd>
 <div class="block">Setter for game strategy</div>
 </dd>
